+++ ChangeLog.md
@@ -2,12 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+
+
+## 2023-01-20
+
+### Added
+- Define package test:  - the behavior of the price rules is checked
+                        - test when the packet should return error
+
+
+
 ## 2023-01-19
 
 ### Added
 
 - Add the remove method to the checkout class & update main.py.
 - Manage the error of receiving wrong product codes.
+
 ## 2023-01-18
 
 ### Added