+++ ChangeLog.md
@@ -3,11 +3,17 @@
 All notable changes to this project will be documented in this file.
 
 
+## 2023-01-21
+
+### Added
+- Add the content to "REAME.md" to explain how to install, test and run the package.
+- Update test
+
 ## 2023-01-20
 
 ### Added
-- Define package test:  - the behavior of the price rules is checked
-                        - test when the packet should return error
+- Define package test:  - the behavior of the price rules is checked.
+                        - test when the packet should return error.
 
 ## 2023-01-19