+++ ChangeLog.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 
 
+
+## 0.0.1 - 2023-01-22
+
+### Added
+- Update README.md
+- Update package
+
 ## 2023-01-21
 
 ### Added